+++ nengo/utils/__init__.py
@@ -11,9 +11,7 @@ from . import logging
 from . import magic
 from . import nco
 from . import network
-from . import neurons
 from . import numpy
 from . import paths
-from . import probe
 from . import simulator
 from . import stdlib